+++ src/content/providers/StringWordProvider.h
@@ -15,8 +15,8 @@ class StringWordProvider : public WordProvider {
   StyledWord getNextWord() override;
   StyledWord getPrevWord() override;
 
-  float getPercentage() override;
-  float getPercentage(int index) override;
+  uint32_t getPercentage() override;
+  uint32_t getPercentage(int index) override;
   void setPosition(int index) override;
   int getCurrentIndex() override;
   char peekChar(int offset = 0) override;
